Add confirm to delete post

+++ src/js/ui/post/delete.ts
@@ -2,11 +2,16 @@ import { deletePost } from "@api/post/delete";
 import { renderPosts } from "@/js/router/views/home";
 
 async function onDeletePost(postId: number) {
-  try {
-    await deletePost(postId);
-    renderPosts();
-  } catch (error) {
-    console.error(error);
+  const confirmed = confirm("Are you sure you want to delete this post?");
+  if (confirmed) {
+    try {
+      await deletePost(postId);
+      renderPosts();
+    } catch (error) {
+      console.error(error);
+    }
+  } else {
+    return;
   }
 }
